Some more fixes
More improvements to the core, getting sub-plugins one step closer to being able to function standalone. Using an ArrayList instead of just a basic array due to the need of a dynamic and global array to handle the sub-plugin ArrayList. Added and extra forward. Minor fix-ups with SourceComms/BaseComms functionality.