So many much lots fixes
Zombie LR crashed me so that's removed for now. VSH sub-plugin finally works like it's supposed to, huzzah. Damage fall-off cvar should work with the new TF2 damagetypes. The plugin is practically stable as of now, but future changes will made and then fixes to fix the changes. It's a grand cycle of things.